+++ src/js/app/screen/travel.js
@@ -31,7 +31,6 @@ app.screen.travel = app.screenManager.invent({
       app.screen.dock.clearFocusMemory()
       app.screen.gameMenu.clearFocusMemory()
       app.screen.ports.clearFocusMemory()
-      app.screen.travel.clearFocusMemory()
 
       this.change('game')
 
@@ -55,7 +54,6 @@ app.screen.travel = app.screenManager.invent({
       app.screen.dock.clearFocusMemory()
       app.screen.gameMenu.clearFocusMemory()
       app.screen.ports.clearFocusMemory()
-      app.screen.travel.clearFocusMemory()
 
       this.change('dock')
     },